DecodeRels docs
