+++ lib/tesla/middleware/decode_rels.ex
@@ -1,0 +1,49 @@
+defmodule Tesla.Middleware.DecodeRels do
+  @behaviour Tesla.Middleware
+
+  @moduledoc """
+  Decode `Link` Hypermedia HTTP header into `opts[:rels]` field in response.
+
+
+  ### Example usage
+  ```
+  defmodule MyClient do
+    use Tesla
+
+    plug Tesla.Middleware.DecodeRels
+  end
+
+  env = MyClient.get("/...")
+  env.opts[:rels] # => %{"Next" => "http://...", "Prev" => "..."}
+  ```
+
+  """
+
+  def call(env, next, _opts) do
+    env
+    |> Tesla.run(next)
+    |> parse_rels
+  end
+
+  defp parse_rels(env) do
+    if link = env.headers["link"] do
+      Tesla.put_opt(env, :rels, rels(link))
+    else
+      env
+    end
+  end
+
+  defp rels(link) do
+    link
+    |> String.split(",")
+    |> Enum.map(&String.trim/1)
+    |> Enum.map(&rel/1)
+    |> Enum.into(%{})
+  end
+
+  defp rel(item) do
+    Regex.run(~r/\A<(.+)>; rel="(.+)"\z/, item, capture: :all_but_first)
+    |> Enum.reverse
+    |> List.to_tuple
+  end
+end
